refactor: make a sub-package of CHAP for each experiment-specific workflow.
