+++ CHAP/__init__.py
@@ -1,0 +1,3 @@
+from CHAP.reader import Reader
+from CHAP.processor import Processor
+from CHAP.writer import Writer
