+++ CHAP/common/models/__init__.py
@@ -1,0 +1,2 @@
+from CHAP.common.models.map import MapConfig
+from CHAP.common.models.integration import IntegrationConfig
