+++ CHAP/edd/README.md
@@ -1,0 +1,2 @@
+# CHAP.edd
+Contains definitions of objects used in a `Pipeline` that operates on data from an EDD experiment.
