+++ CHAP/edd/__init__.py
@@ -1,0 +1,7 @@
+# from CHAP.edd.reader import
+from CHAP.edd.processor import (MCACeriaCalibrationProcessor,
+                                MCADataProcessor)
+# from CHAP.edd.writer import
+
+from CHAP.common import (MapProcessor,
+                         StrainAnalysisProcessor)
