+++ CHAP/inference/README.md
@@ -1,0 +1,2 @@
+# CHAP.inference
+Contains definitions of objects used for interacting with a [TFaaS](https://github.com/vkuznet/TFaaS) server in a `Pipeline`.
