+++ CHAP/inference/__init__.py
@@ -1,0 +1,3 @@
+# from CHAP.inference.reader import
+from CHAP.inference.processor import TFaaSImageProcessor
+# from CHAP.inference.writer import
