+++ CHAP/saxswaxs/README.md
@@ -1,0 +1,2 @@
+# CHAP.saxswaxs
+Contains definitions of objects used in a `Pipeline` that operates on data from a SAXS/WAXS experiment.
