+++ CHAP/saxswaxs/__init__.py
@@ -1,0 +1,6 @@
+# from CHAP.saxswaxs.reader import
+# from CHAP.saxswaxs.processor import
+# from CHAP.saxswaxs.writer import
+
+from CHAP.common import (IntegrateMapProcessor,
+                         MapProcessor)
