+++ CHAP/saxswaxs/processor.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python  
+
+if __name__ == '__main__':
+    from CHAP.processor import main
+    main()
