+++ CHAP/sin2psi/README.md
@@ -1,0 +1,2 @@
+# CHAP.sin2psi
+Contains definitions of objects used in a `Pipeline` that operates on data from a sin2psi experiment.
