+++ CHAP/sin2psi/__init__.py
@@ -1,0 +1,7 @@
+# from CHAP.sin2psi.reader import
+# from CHAP.sin2psi.processor import
+# from CHAP.sin2psi.writer import
+
+from CHAP.common import (IntegrateMapProcessor,
+                         MapProcessor,
+                         StrainAnalysisProcessor)
