+++ CHAP/sin2psi/writer.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python  
+
+if __name__ == '__main__':
+    from CHAP.writer import main
+    main()
