+++ test-data/config.yaml
@@ -2,5 +2,6 @@ pipeline:
   - reader.Reader:
       filename: data.csv
   - processor.Processor: {}
+  - common.PrintProcessor: {}
   - writer.Writer:
       filename: data.out
